refactored to get the linting to work
